{"schema":"decodex/commit/1","summary":"Split quantitative audit manifests","authority":"manual"}
